third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/html/rendering/widgets/button-layout/abspos-expected.txt

This is a testharness.js-based test.
FAIL abspos button with auto width, non-auto left/right (rtl) assert_equals: offsetLeft expected 665 but got 100
PASS abspos button with auto width, non-auto left/right (ltr)
Harness: the test ran to completion.
